% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createSequence}
\alias{createSequence}
\title{Create a sequence from toi phenx}
\usage{
createSequence(firstPhenx, secondPhenx, phenxLength = 7L)
}
\arguments{
\item{firstPhenx}{Integer.}

\item{secondPhenx}{Integer.}

\item{phenxLength}{Integer, number of digits used to represent the second phenx in the sequence.}
}
\value{
Integer.
}
\description{
Function create a sequence from toi phenx.
}
