% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createTransitiveSequences}
\alias{createTransitiveSequences}
\title{create all transitive Sequences}
\usage{
createTransitiveSequences(
  df_dbMart,
  outputDir,
  outputFilePrefix,
  numOfThreads = 1L
)
}
\arguments{
\item{df_dbMart}{The dataframe that stores the numeric data mart.}

\item{outputDir}{The path as string to the directory where the sequences should be stored.}

\item{outputFilePrefix}{The string file prefix for the patient files storing the sequences.}

\item{numOfThreads}{The number of threads that should be used during sequencing.}
}
\value{
the overall number of sequences stored.
}
\description{
Function to create all transitive sequences
an store in binary format them to files.
}
