\name{extractAllTransiviteSequences}
\alias{extractAllTransiviteSequences(df_dbMart, outputDir, outputFilePrefix, numOfThreads,
                   addDuration = TRUE, durationInWeeks = FALSE,
                   durationInMonths = FALSE)}
\docType{package}
\title{
Simple function using Rcpp
}
\description{
Simple function using Rcpp
}
\usage{
extractAllTransiviteSequences(df_dbMart, outputDir, outputFilePrefix, numOfThreads,
                   addDuration = TRUE, durationInWeeks = FALSE,
                   durationInMonths = FALSE)	
}
\examples{
\dontrun{
extractAllTransiviteSequences(df_dbMart, outputDir, outputFilePrefix, numOfThreads,
                   addDuration = TRUE, durationInWeeks = TRUE,
                   durationInMonths = TRUE)
}


}
