\name{extractNonSparseSequences}
\alias{extractNonSparseSequences}
\docType{package}
\title{
create transitive Sequences from EHR
}
\description{
The function extracts sequences from given DBMart and returns the sequences as a data frame.
}
\usage{
extractNonSparseSequences()	
}
\examples{
\dontrun{
extractNonSparseSequences()
}
}
