% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEndPhenxFromSequence}
\alias{getEndPhenxFromSequence}
\title{Get End Phenx From Sequence}
\usage{
getEndPhenxFromSequence(sequence, phenxLength = 7L)
}
\arguments{
\item{sequence}{Integer.}

\item{phenxLength}{Integer, number of digits used to represent the second phenx in the sequence.}
}
\value{
The endPhenx of a sequence.
}
\description{
Function to extract the end Phenx from a sequence.
}
