% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStartPhenxFromSequence}
\alias{getStartPhenxFromSequence}
\title{Get Start Phenx From Sequence}
\usage{
getStartPhenxFromSequence(sequence, phenxLength = 7L)
}
\arguments{
\item{sequence}{Integer.}

\item{phenxLength}{Integer, number of digits used to represent the second phenx in the sequence.}
}
\value{
The startPhenx of a sequence
}
\description{
Function to extract the start Phenx from a sequence.
}
