% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sequenceAndSummarize}
\alias{sequenceAndSummarize}
\title{Sequences and summarizes all sequences in a dbmart}
\usage{
sequenceAndSummarize(
  df_dbMart,
  lowerBucketThresholds,
  storeSeqDuringCreation = FALSE,
  includeDurations = FALSE,
  numOfThreads = 1L,
  outputDir = "",
  outputFilePrefix = "",
  removeSparseSequences = TRUE,
  sparsityValue = 0.05,
  createTemporalBuckets = FALSE,
  durationSparsity = FALSE,
  durationSparsityValue = 0,
  removeSparseTemporalBuckets = FALSE,
  patIdLength = 7L,
  durationPeriods = 30.437,
  daysForCoOoccurence = 14L,
  summaryOnPatientLevel = FALSE
)
}
\arguments{
\item{df_dbMart}{The data frame that stores the data mart.}

\item{lowerBucketThresholds}{IntegerVector, lower duration Thresholds for the duration buckets of the candidate sequences}

\item{storeSeqDuringCreation}{Boolean parameter to control if the duration should be included in the sequence ID during creation, DEFAULT = FALSE.}

\item{includeDurations}{include duration buckets in the summary}

\item{numOfThreads}{The number of threads that should be used during sequencing.}

\item{outputDir}{The path as string to the directory where the sequences should be stored.}

\item{outputFilePrefix}{The string file prefix for the patient files storing the sequences.}

\item{removeSparseSequences}{Boolean parameter to control if the sparsity should be applied.}

\item{sparsityValue}{The numeric value for the sparsity. DEFAULT = 0.05.}

\item{createTemporalBuckets}{Boolean flag if the the the sequences should be split up in dynamic buckets. Number of buckets min(4, max_duration(sequence)).}

\item{durationSparsity}{Boolean flag to control if sparse sequences should be removed considering the duration periods of a sequence.}

\item{durationSparsityValue}{Numeric value.}

\item{removeSparseTemporalBuckets}{Boolean, to control if the sparsity should be applied on the dynamic temporal buckets.}

\item{patIdLength}{Integer, describes the number of digits that are used for the patient number.}

\item{durationPeriods}{Numeric, Upper threshold, stores the number of day in the time period, e.g. 30.471 for months, 364.25 for years.}

\item{daysForCoOoccurence}{Integer, sets the upper threshold for the sequence duration so that they are counted as co-occurrence (meaning a duration of 0).}

\item{summaryOnPatientLevel}{bool, that defines if the summary should be on the patient level (counting occurrences for each patient) or on the dbMart level}
}
\value{
The summary as data frame.
}
\description{
Summarizes the sequences in a dbmart
}
