% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize}
\alias{summarize}
\title{summarize}
\arguments{
\item{numOfThreads}{The number of threads that should be used during sequencing.}

\item{lowerBucketThresholds}{IntegerVector, lower duration Thresholds for the duration buckets of the candidate sequences}

\item{includeDurations}{include duration buckets in the summary}

\item{sequences}{the vector containing the sequences to summarize}
}
\value{
The summary as data frame.
}
\description{
Function to summarize a set of sequences and  returing a R data frame that contains the summary.
that calls this functions with predefined parameters.
}
