% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transformDataFrameToStruct}
\alias{transformDataFrameToStruct}
\title{transform the dbmart data frame to c++ structure
Function to create all transitive sequences
an store in binary format them to files.}
\arguments{
\item{df_dbMart}{The dataframe that stores the numeric data mart.}
}
\value{
dbMart as c++ struct
}
\description{
transform the dbmart data frame to c++ structure
Function to create all transitive sequences
an store in binary format them to files.
}
