\name{transformDbMartToNumeric}
\alias{transformDbMartToNumeric}
\docType{package}
\title{
Function to convert a alpha-numeric dbMArt to numeric.
}
\description{
Function to converts a dbMart with alpha-numeric patient IDs and phenx to numeric IDs and returns the transformed dbMArt and look up tables for the numeric patient ID and phenx.
}
\usage{
transformDbMartToNumeric()	
}

