% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transformToDefaultDataFrame}
\alias{transformToDefaultDataFrame}
\title{Transforms the C++ representation used to store the sequences into R}
\arguments{
\item{sequences}{std::vector of the sequences that should be transformed into a R dataframe}

\item{returnDuration}{Boolean, if the duration should be included into the returned dataframe}
}
\description{
Function to transform a vector of the C++ representation used to store the sequences into an R dataframe.
}
