% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tSPMPlus}
\alias{tSPMPlus}
\title{tSPMPlus}
\usage{
tSPMPlus(
  df_dbMart,
  storeSeqDuringCreation = FALSE,
  outputDir = "",
  outputFilePrefix = "",
  numOfThreads = 1L,
  removeSparseSequences = TRUE,
  sparsityValue = 0.05,
  createTemporalBuckets = FALSE,
  durationSparsity = FALSE,
  durationSparsityValue = 0,
  removeSparseTemporalBuckets = FALSE,
  patIdLength = 7L,
  returnDuration = TRUE,
  durationPeriods = 30.437,
  daysForCoOoccurence = 14L
)
}
\arguments{
\item{df_dbMart}{The data frame that stores the data mart.}

\item{storeSeqDuringCreation}{Boolean parameter to control if the duration should be included in the sequence ID during creation, DEFAULT = FALSE.}

\item{outputDir}{The path as string to the directory where the sequences should be stored.}

\item{outputFilePrefix}{The string file prefix for the patient files storing the sequences.}

\item{numOfThreads}{The number of threads that should be used during sequencing.}

\item{removeSparseSequences}{Boolean parameter to control if the sparsity should be applied}

\item{sparsityValue}{The numeric value for the sparsity. DEFAULT = 0.05.}

\item{createTemporalBuckets}{Boolean flag if the the the sequences should be split up in dynamic buckets. Number of buckets min(4, max_duration(sequence)).}

\item{durationSparsity}{Boolean flag to control if sparse sequences should be removed considering the duration periods of a sequence.}

\item{durationSparsityValue}{Numeric value.}

\item{removeSparseTemporalBuckets}{Boolean, to control if the sparsity should be applied on the dynamic temporal buckets.}

\item{patIdLength}{describes the number of digits that are used for the patient number.}

\item{returnDuration}{Boolean, controls if the data frame that is returns contains.}

\item{durationPeriods}{Numeric, Upper threshold, stores the number of day in the time period, e.g. 30.471 for months, 364.25 for years.}

\item{daysForCoOoccurence}{Integer, sets the upper threshold for the sequence duration so that they are counted as co-occurrence (meaning a duration of 0).}
}
\value{
The sequences as data frame.
}
\description{
Function to call the tSPM+ workflow, most of the other function provide by this package are wrapper functions
that calls this functions with predefined parameters.
}
