\name{tSPMPlusR-package}
\alias{tSPMPlusR-package}
\alias{tSPMPlusR}
\docType{package}
\title{
  transitive Sequence Pattern Mining for EHR records 
}
\description{
  R package with C++ backend for fast, low memory consuming creation
  of transitive sequences. The package provides additional functionality to add the temportal lenght of the sequences.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jonas Huegel, Hossein Estiri

Maintainer: jhuegel@mgh.harvard.edu
}
\references{
  tSPM algorithm:
}
\keyword{ tSPM, MLHO, temporal sequencing }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
