% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractAllTransiviteSequences}
\alias{extractAllTransiviteSequences}
\title{Extract All Transivite Sequences}
\usage{
extractAllTransiviteSequences(
  df_dbMart,
  storeSeqDuringCreation = FALSE,
  outputDir = "",
  outputFilePrefix = "",
  numOfThreads = 1L,
  returnDuration = TRUE,
  durationPeriods = 30.437,
  daysForCoOoccurence = 14L
)
}
\arguments{
\item{df_dbMart}{The data frame that stores the data mart.}

\item{storeSeqDuringCreation}{Boolean parameter to control if the duration should be included in the sequence ID during creation, DEFAULT = FALSE.}

\item{outputDir}{The path as string to the directory where the sequences should be stored.}

\item{outputFilePrefix}{The string file prefix for the patient files storing the sequences.}

\item{numOfThreads}{The number of threads that should be used during sequencing.}

\item{returnDuration}{Boolean, controls if the data frame that is returns contains.}

\item{durationPeriods}{Numeric, Upper threshold, stores the number of day in the time period, e.g. 30.471 for months, 364.25 for years.}

\item{daysForCoOoccurence}{Integer, sets the upper threshold for the sequence duration so that they are counted as co-occurrence (meaning a duration of 0).}
}
\value{
The sequences as data frame.
}
\description{
Function to extract all transitive sequences.
}
