% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformDbMartToNumeric}
\alias{transformDbMartToNumeric}
\title{Transforms an alphanumeric data frame to a numeric one}
\usage{
transformDbMartToNumeric(dbmart)
}
\arguments{
\item{dbmart}{The alpha numeric dbmart that should be converted.}
}
\value{
A list of data frames containing the numeric datamart and the patient id and phenx look up tables.
}
\description{
Transforms an alphanumeric data frame to a numeric one and returns it together with the look up tables for the patient id and phenx.
}
